+++ branches/multiple-layers-branch/qgis/providers/postgres/qgspostgresprovider.h
@@ -186,7 +186,7 @@ class QgsPostgresProvider:public QgsVectorDataProvider
 
     /** Return the extent for this data layer
     */
-    virtual QgsRect *extent(int dataSourceLayerNum = 0);
+    virtual QgsRect *extent(size_t dataSourceLayerNum = 0);
 
     /**
      * Get the attributes associated with a feature
